initial version
timeMatrices exploded : the size of the timeMatrix is now equal to (numberOfServices*numberOfServices)

functions : IdIndex() and IndexId() usefull for relation constraints and for displaying results

timesFromWarehouse : vector of maps, mapping the distance from a start_index to each service.

timesFromWarehouse : vector of maps, mapping the distance from each service to an end_index.

implementation of calculation of route durations with timeFromWarehouse and timeToWarehouse.
 Adding duration constraints, adding totalDuration (= sum of each route Duration) to the objective.

Implementation of capacity constraints : for classic VRP and for PickUp and Delivery.

Implementation of precedence constraints for some relations (Pickup and delivery and Order)

Implementation of unassigned services with an extra super vehicle : calculation of exclusionCost to minimize.